requirements.txt
scikit-learn
wandb
opencv-python
numpy
pillow

torch
torchvision

git+https://github.com/NVIDIA/apex.git@master#egg=apex
